requirements.txt
# --------------------------------------------------------------------
# 🌌 HelixNet Core API Dependencies
# --------------------------------------------------------------------
# APPLICATION FRAMEWORK & SERVER
fastapi==0.111.0
uvicorn[standard]==0.29.0
databases
python-multipart  # For form data
# DATABASE STACK (PostgreSQL/Async)
sqlalchemy==2.0.30
sqlalchemy-utils==0.41.2
alembic==1.13.1
asyncpg==0.29.0
psycopg[binary]
# DATA MODELS & CONFIGURATION
pydantic==2.7.4
pydantic-settings==2.11.0
python-dotenv
pyyaml
# SECURITY & AUTHENTICATION (The Chuck Norris Security Core)
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.1,<4.1
cryptography==41.0.4
fastapi-cors
python-jose
# ASYNCHRONOUS TASK QUEUE (Celery)
celery==5.4.0
redis==5.0.4
pika==1.2.0
flower==2.0.1
# UTILITIES & STORAGE
jinja2
requests
aiofiles
minio==7.1.3  # For S3/MinIO storage
loguru==0.7.0
sentry-sdk==1.28.0
# --------------------------------------------------------------------
# 🛠️ DEVELOPMENT & TESTING
# --------------------------------------------------------------------
pytest==7.4.2
httpx==0.24.1
pytest-asyncio==0.22.0
pytest-cov==4.1.0
ipython
black
isort
flake8
mypy
pre-commit